#94 refactor: forum logs processing
